+++ source/scener/content/gltf/buffer_view.cpp
@@ -7,12 +7,12 @@
 
 namespace scener::content::gltf
 {
-    std::size_t buffer_view::byte_offset() const noexcept
+    std::uint32_t buffer_view::byte_offset() const noexcept
     {
         return _byte_offset;
     }
 
-    std::size_t buffer_view::byte_length() const noexcept
+    std::uint32_t buffer_view::byte_length() const noexcept
     {
         return _byte_length;
     }
@@ -27,7 +27,7 @@ namespace scener::content::gltf
         return get_data(0, _byte_length);
     }
 
-    gsl::span<const std::uint8_t> buffer_view::get_data(std::size_t offset, std::size_t count) const noexcept
+    gsl::span<const std::uint8_t> buffer_view::get_data(std::uint32_t offset, std::uint32_t count) const noexcept
     {
         return _buffer->get_data(_byte_offset + offset, count);
     }
