+++ source/scener/graphics/texture.hpp
@@ -37,7 +37,7 @@ namespace scener::graphics
 
         /// Gets the number of texture levels in a multilevel texture.
         /// \returns the number of texture levels in a multilevel texture.
-        virtual std::size_t level_count() const noexcept = 0;
+        virtual std::uint32_t level_count() const noexcept = 0;
 
         /// Activates the texture object.
         virtual void bind() const noexcept = 0;
